+++ src/main/kotlin/icu/windea/pls/cwt/config/CwtHardCodedConfig.kt
@@ -1,4 +1,0 @@
-package icu.windea.pls.cwt.config
-
-interface CwtHardCodedConfig
-