+++ src/main/kotlin/icu/windea/pls/cwt/config/CwtModifierCategoryConfig.kt
@@ -7,9 +7,10 @@ import icu.windea.pls.cwt.psi.*
  * @property internalId internal_id: int
  * @property supportedScopes supported_scopes: string[]
  */
-data class CwtModifyCategoryConfig(
+data class CwtModifierCategoryConfig(
 	override val pointer: SmartPsiElementPointer<CwtProperty>,
 	val name:String,
-	val internalId: Int,
+	val internalId: String,
 	val supportedScopes:List<String>
 ): CwtConfig<CwtProperty>
+