+++ src/main/kotlin/icu/windea/pls/cwt/config/CwtModifierDefinitionConfig.kt
@@ -1,6 +1,0 @@
-package icu.windea.pls.cwt.config
-
-data class CwtModifierDefinitionConfig(
-	val tag:String,
-	val categories:String //string | int[2^?]
-): CwtHardCodedConfig